requirements.txt
# 核心依赖
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.27.0
peft>=0.8.0
bitsandbytes>=0.41.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0

# 可视化和 Web
gradio>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 工具
tqdm>=4.65.0
pyyaml>=6.0
datasets>=2.14.0

# LLaMA-Factory (通过 git clone 安装)
# git clone https://github.com/hiyouga/LLaMA-Factory.git
# cd LLaMA-Factory && pip install -e .
